Fix recursive import error
